requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
black==24.4.2
blinker==1.8.2
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.7
django-celery-beat==2.6.0
django-debug-toolbar==4.4.6
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.1.0
flake8-annotations==3.1.1
flake8-quotes==3.4.0
flake8-variables-names==0.0.6
flower==2.0.1
humanize==4.10.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg==3.2.1
psycopg-binary==3.2.1
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.3.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
setuptools==71.1.0
six==1.16.0
sqlparse==0.5.1
stripe==10.3.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
